refactor: UserFactoryの作成
